requirements-dev.txt
# Developer + optional extras
-r requirements.txt

pytest>=8.0
ruff>=0.4
mypy>=1.8

# API (optional)
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.0
python-multipart>=0.0.9

# OCR (optional)
pillow>=10.0
pytesseract>=0.3.10
pdf2image>=1.17.0
